Add support for inspecting serialized tokens
Previously, serialized tokens were parsed, and then recreated with a static
root key, so that we could run an authorizer without providing a public key.
This is not possible any more with 3rd party blocks, since recreating the
token would require the external secret keys.

Instead, we ask for the public key and directly inspect the serialized token.