execute: parse initial block values
execute delegates to execute_serialized by crafting a token.
execute_serialized highlights checks by parsing the contents of the
token. The issue was that `execute` was still displaying the actual
code provided by the user, which was not the same as the code generated
from a serialized token (comments are stripped, amongst other things).
So the solution was to make execute_serialize take an optional vec of
original source code, and only parse code from the token if nothing else
is available